8-hex-digit seeds: 32-bit seed override + title editor
Widen the seed to 8 hex digits (g_seed_override / g_run_seed are now
uint32). The title SEED editor edits all eight digits; the run seed is
shown 8-wide on the game-over screen and pins/reproduces a run.

The RNG state stays 16-bit (fast, bank-safe, all maps unchanged), so the
32-bit seed is XOR-folded into it (seed16 = lo ^ hi). A 16-bit-only seed
like the debug pin 0x00001D2D folds to itself, so existing maps/tests are
untouched. Effective distinct-map space is still 2^16, but every 8-digit
code reproduces the same run.

Verified: PyBoy confirms an 8-digit seed (0xDEADBEEF) sets g_run_seed and
shows on the game-over screen; the editor shows 8 digits; suite 13/13.

+++ src/world.c
@@ -23,8 +23,8 @@ uint8_t g_debug;         /* set at the title; not persisted */
    world_new). g_seed_override != 0 forces that seed for the next new
    game; g_run_seed records the seed actually used so it can be shown and
    reproduced. */
-uint16_t g_seed_override;
-uint16_t g_run_seed;
+uint32_t g_seed_override;
+uint32_t g_run_seed;
 uint8_t  g_repeat_speed = 1;
 uint16_t g_wander_t;
 uint8_t  g_lang = 1;     /* player setting — survives new games;
